refactor; rename spacy_tokenizer -> stream_spacy_tokenizer
